fix(test): bless profile catalog fixture and resolve Kotlin warnings
The governance test failed because the bundled asset gained QUIC
targets that were not reflected in the contract fixture. Also fix
7 Kotlin compiler warnings: always-true instance checks, nullable
receiver, nullable String from System.getProperty, and redundant
toInt() conversion.

+++ core/service/src/test/kotlin/com/poyka/ripdpi/services/NetworkFingerprintMapperTest.kt
@@ -22,7 +22,7 @@ class NetworkFingerprintMapperTest {
                         CapturedWifiIdentity(
                             ssid = "\"Cafe Wifi\"",
                             bssid = "AA:BB:CC:DD:EE:FF",
-                            gatewayIpv4 = 0x0101A8C0.toInt(),
+                            gatewayIpv4 = 0x0101A8C0,
                         ),
                     metered = true,
                 ),
